CSP-81 Moved and renamed GetAllProviderStandards handler
