+++ src/SFA.DAS.Roatp.CourseManagement.Application.UnitTests/Handlers/GetStandardQueryHandlerTests.cs
@@ -3,7 +3,7 @@ using FluentAssertions;
 using Microsoft.Extensions.Logging;
 using Moq;
 using NUnit.Framework;
-using SFA.DAS.Roatp.CourseManagement.Application.Standards.Queries;
+using SFA.DAS.Roatp.CourseManagement.Application.ProviderStandards.Queries.GetAllProviderStandards;
 using SFA.DAS.Roatp.CourseManagement.Domain.Interfaces;
 using System;
 using System.Collections.Generic;
@@ -14,11 +14,11 @@ namespace SFA.DAS.Roatp.CourseManagement.Application.UnitTests.Handlers
 {
     public class GetStandardQueryHandlerTests
     {
-        private GetStandardQueryHandler _handler;
+        private GetAllProviderStandardsQueryHandler _handler;
         private Mock<IApiClient> _apiClient;
-        private Mock<ILogger<GetStandardQueryHandler>> _logger;
-        private GetStandardQuery _query;
-        private GetStandardQueryResult _queryResult;
+        private Mock<ILogger<GetAllProviderStandardsQueryHandler>> _logger;
+        private GetAllProviderStandardsQuery _query;
+        private GetAllProviderStandardsQueryResult _queryResult;
         private List<Domain.ApiModels.Standard> _standards;
 
         [SetUp]
@@ -26,19 +26,19 @@ namespace SFA.DAS.Roatp.CourseManagement.Application.UnitTests.Handlers
         {
             var autoFixture = new Fixture();
 
-            _query = autoFixture.Create<GetStandardQuery>();
-            _queryResult = autoFixture.Create<GetStandardQueryResult>();
+            _query = autoFixture.Create<GetAllProviderStandardsQuery>();
+            _queryResult = autoFixture.Create<GetAllProviderStandardsQueryResult>();
             _standards = autoFixture.Create<List<Domain.ApiModels.Standard>>();
             _queryResult.Standards = _standards;
             _apiClient = new Mock<IApiClient>();
-            _logger = new Mock<ILogger<GetStandardQueryHandler>>();
+            _logger = new Mock<ILogger<GetAllProviderStandardsQueryHandler>>();
         }
 
         [Test]
         public async Task Handle_ValidRequest_ReturnsValidResponse()
         {
             _apiClient.Setup(x => x.Get<List<Domain.ApiModels.Standard>>($"Standards/{_query.Ukprn}")).ReturnsAsync(() => _standards);
-            _handler = new GetStandardQueryHandler(_apiClient.Object, _logger.Object);
+            _handler = new GetAllProviderStandardsQueryHandler(_apiClient.Object, _logger.Object);
 
             var result = await _handler.Handle(_query, CancellationToken.None);
             result.Should().NotBeNull();
@@ -49,7 +49,7 @@ namespace SFA.DAS.Roatp.CourseManagement.Application.UnitTests.Handlers
         public async Task Handle_NoStandardsReturned_ReturnsNullResponse()
         {
             _apiClient.Setup(x => x.Get<List<Domain.ApiModels.Standard>>($"Standards/{_query.Ukprn}")).ReturnsAsync(() => null);
-            _handler = new GetStandardQueryHandler(_apiClient.Object, _logger.Object);
+            _handler = new GetAllProviderStandardsQueryHandler(_apiClient.Object, _logger.Object);
 
             var result = await _handler.Handle(_query, CancellationToken.None);
 
@@ -61,7 +61,7 @@ namespace SFA.DAS.Roatp.CourseManagement.Application.UnitTests.Handlers
         public void GetStandardQueryHandler_Returns_Exception()
         {
             _apiClient.Setup(x => x.Get<List<Domain.ApiModels.Standard>>($"Standards/{_query.Ukprn}")).Throws(new Exception());
-            _handler = new GetStandardQueryHandler(_apiClient.Object, _logger.Object);
+            _handler = new GetAllProviderStandardsQueryHandler(_apiClient.Object, _logger.Object);
             Assert.ThrowsAsync<Exception>(() => _handler.Handle(_query, CancellationToken.None));
         }
     }
