+++ src/SFA.DAS.Roatp.CourseManagement.Application/ProviderStandards/Queries/GetAllProviderStandards/GetAllProviderStandardsQuery.cs
@@ -1,0 +1,14 @@
+﻿using MediatR;
+
+namespace SFA.DAS.Roatp.CourseManagement.Application.ProviderStandards.Queries.GetAllProviderStandards
+{
+    public class GetAllProviderStandardsQuery : IRequest<GetAllProviderStandardsQueryResult>
+    {
+        public int Ukprn { get; }
+
+        public GetAllProviderStandardsQuery(int ukprn)
+        {
+            Ukprn = ukprn;
+        }
+    }
+}
