+++ src/SFA.DAS.Roatp.CourseManagement.Application/ProviderStandards/Queries/GetAllProviderStandards/GetAllProviderStandardsQueryResult.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace SFA.DAS.Roatp.CourseManagement.Application.ProviderStandards.Queries.GetAllProviderStandards
+{
+    public class GetAllProviderStandardsQueryResult
+    {
+        public List<Domain.ApiModels.Standard> Standards { get; set; }
+    }
+}
