+++ src/SFA.DAS.Roatp.CourseManagement.Application/SFA.DAS.Roatp.CourseManagement.Application.csproj
@@ -13,4 +13,8 @@
     <ProjectReference Include="..\SFA.DAS.Roatp.CourseManagement.Domain\SFA.DAS.Roatp.CourseManagement.Domain.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Standards\Queries\" />
+  </ItemGroup>
+
 </Project>
