--- src/SFA.DAS.Roatp.CourseManagement.Application/Standards/Queries/GetStandardQuery.cs
@@ -1,14 +1,0 @@
-﻿using MediatR;
-
-namespace SFA.DAS.Roatp.CourseManagement.Application.Standards.Queries
-{
-    public class GetStandardQuery : IRequest<GetStandardQueryResult>
-    {
-        public int Ukprn { get; }
-
-        public GetStandardQuery(int ukprn)
-        {
-            Ukprn = ukprn;
-        }
-    }
-}
