--- src/SFA.DAS.Roatp.CourseManagement.Application/Standards/Queries/GetStandardQueryResult.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace SFA.DAS.Roatp.CourseManagement.Application.Standards.Queries
-{
-    public class GetStandardQueryResult
-    {
-        public List<Domain.ApiModels.Standard> Standards { get; set; }
-    }
-}
