+++ src/SFA.DAS.Roatp.CourseManagement.Web/Controllers/StandardsController.cs
@@ -2,8 +2,8 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using SFA.DAS.Roatp.CourseManagement.Application.ProviderStandards.Queries.GetAllProviderStandards;
 using SFA.DAS.Roatp.CourseManagement.Application.Standard.Queries;
-using SFA.DAS.Roatp.CourseManagement.Application.Standards.Queries;
 using SFA.DAS.Roatp.CourseManagement.Web.Infrastructure;
 using SFA.DAS.Roatp.CourseManagement.Web.Infrastructure.Authorization;
 using SFA.DAS.Roatp.CourseManagement.Web.Models.Standards;
@@ -30,7 +30,7 @@ namespace SFA.DAS.Roatp.CourseManagement.Web.Controllers
         {
             _logger.LogInformation("Getting standards for {ukprn}", Ukprn);
 
-            var result = await _mediator.Send(new GetStandardQuery(Ukprn));
+            var result = await _mediator.Send(new GetAllProviderStandardsQuery(Ukprn));
 
             var model = new StandardListViewModel
             {
