+++ src/SFA.DAS.Roatp.CourseManagement.Web/Startup.cs
@@ -12,8 +12,8 @@ using Microsoft.Extensions.Hosting;
 using SFA.DAS.Configuration.AzureTableStorage;
 using SFA.DAS.Provider.Shared.UI;
 using SFA.DAS.Provider.Shared.UI.Startup;
+using SFA.DAS.Roatp.CourseManagement.Application.ProviderStandards.Queries.GetAllProviderStandards;
 using SFA.DAS.Roatp.CourseManagement.Application.Standards.Commands.UpdateContactDetails;
-using SFA.DAS.Roatp.CourseManagement.Application.Standards.Queries;
 using SFA.DAS.Roatp.CourseManagement.Domain.Configuration;
 using SFA.DAS.Roatp.CourseManagement.Domain.Interfaces;
 using SFA.DAS.Roatp.CourseManagement.Infrastructure.ApiClients;
@@ -74,7 +74,7 @@ namespace SFA.DAS.Roatp.CourseManagement.Web
 
             services.AddProviderUiServiceRegistration(_configuration);
 
-            services.AddMediatR(typeof(GetStandardQueryHandler).Assembly);
+            services.AddMediatR(typeof(GetAllProviderStandardsQueryHandler).Assembly);
 
             services.AddAuthorizationServicePolicies();
